requirements.txt
pandas
streamlit
streamlit-agraph
streamlit-tree-select
openai>=1.0.0
python-dotenv